Move Dialogs to componentes package
