+++ AdministradorProyectosTP/src/ui/KanbanPanel.java
@@ -7,7 +7,7 @@ import service.ServiceException;
 import javax.swing.*;
 import java.awt.*;
 import java.util.List;
-import ui.Dialogs;
+import ui.componentes.Dialogs;
 
 public class KanbanPanel extends JPanel {
     private final AppManager manager;
